+++ bu-ussd/utils/menuOptions.js
@@ -1,0 +1,23 @@
+module.exports = [
+    {
+        'registration': {
+            'en_bu': 'Register New Client',
+            'bu': 'Kwandika umunywanyi mushaha',
+            'fr_bu': 'Enregistrement d\'un nouveau client'
+        },
+    },
+    {
+        'place_order': {
+            'en_bu': 'Place Order',
+            'bu': 'Gusaba inyongeramwimbu',
+            'fr_bu': 'Commander'
+        },
+    },
+    {
+        'get_balance': {
+            'en_bu': 'Get Balance',
+            'bu': 'Yose hamwe',
+            'fr_bu': 'Avoir la solde'
+        }
+    }
+];